Correctly set/unset favorited in case of error
Only update the flag if the API call did not fail, and update the
button’s status with the actual flag value instead of assuming it was
always updated.

This fixes an issue where, if there’s an error when liking a tweet,
Cawbird still thinks you’ve successfully liked the tweet, and to
actually like it, you have to un-like it (prompting another error, “no
status found with that ID”, because Cawbird attempts to destroy a status
that, server-side, was never created in the first place), and then like
it again. Instead, it now looks like you haven’t liked the tweet yet,
and you can retry immediately without first unliking.